Moved game logic to it's own library
