+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 resolver = "2"
-members = [ "circus", "clown" ]
+members = [ "circus", "clown" , "shitface"]
