+++ shitface/Cargo.toml
@@ -1,10 +1,12 @@
 [package]
-name = "circus"
+name = "shitface"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-clap = { version = "4.5.4", features = ["derive"] }
-shitface = { path = "../shitface" }
+anyhow = "1.0.83"
+itertools = "0.12.1"
+rand = "0.8.5"
+strum = { version = "0.26.2", features = ["derive"] }
