+++ shitface/src/cards/card_suits.rs
@@ -1,0 +1,9 @@
+use strum::EnumIter;
+
+#[derive(Debug, EnumIter, Clone, Copy)]
+pub enum CardSuits {
+    Clubs,
+    Diamonds,
+    Hearts,
+    Spades,
+}
