+++ shitface/src/cards/card_type.rs
@@ -1,0 +1,11 @@
+use strum::EnumIter;
+
+#[derive(Debug, EnumIter, Clone, Copy, PartialEq, Eq)]
+pub enum CardType {
+    Ace,
+    Numeric(u8),
+    Jack,
+    Queen,
+    King,
+    Joker,
+}
