+++ shitface/src/cards/mod.rs
@@ -1,0 +1,4 @@
+pub mod card;
+pub mod card_color;
+pub mod card_suits;
+pub mod card_type;
